chore: move files around and add jsdoc
